Introduced a new zheat_v2 function.
The original zheat function is breaking when length(con) is not equal to one or length(levels(conditions)). This is specially important when length(levels(conditions)) is larger than 2 and an arbitray order of plot columns is desired. Moreover, zheat function currently is sorting the heatmap rows (genes) only based on the first member of "con" vector.

In the new version of the function, the "con" variable is replaced by the new "order.by" variable. If provided the user is obligated to provide a vector of an arbitrary permutation of levels(conditions), otherwise, it would be equal to levels(conditions) by default. Moreover, a new "sort.genes.by" variable is introduced to sort the genes based on multiple conditions.